refactor: extract Polars row_number helper, fix docstring, clean up tests
- Extract _row_number_nulls_last() in PolarsLazyRank to deduplicate
  the null-handling pattern across row_number, ntile, top, and bottom
- Rename "Integer rank types" to "Numeric rank types" (percent_rank
  returns float)
- Add config-based test for top_N (test_option_based_top_n)
- Drop redundant test_top_n_null_order_by and test_bottom_n_null_order_by
  (already covered by full-array assertions in test_top_n_ranked and
  test_bottom_n_ranked)

+++ mloda/community/feature_groups/data_operations/row_preserving/rank/base.py
@@ -19,7 +19,7 @@ class RankFeatureGroup(FeatureChainParserMixin, FeatureGroup):
 
     ## Supported Rank Types
 
-    **Integer rank types:**
+    **Numeric rank types:**
 
     - ``row_number``: Sequential position (1, 2, 3, ...), no ties
     - ``rank``: Standard rank with gaps for ties (1, 2, 2, 4, ...)
